refactor: rpc only checks and selects once at session start
