+++ static/index.html
@@ -178,6 +178,6 @@
     </main>
 
     <ul class="notifications"></ul>
-    <script src="out/scripts/rewards/index.js" type="application/javascript"></script>
+    <script src="out/scripts/rewards/init.js" type="application/javascript"></script>
   </body>
 </html>
