+++ static/scripts/rewards/render-transaction/setPagination.ts
@@ -1,0 +1,12 @@
+import { app } from "../app-state";
+
+export function setPagination(nextTxButton: Element | null, prevTxButton: Element | null) {
+  if (!nextTxButton || !prevTxButton) return;
+  if (app.claims.length > 1) {
+    prevTxButton.classList.remove("hide-pagination");
+    nextTxButton.classList.remove("hide-pagination");
+
+    prevTxButton.classList.add("show-pagination");
+    nextTxButton.classList.add("show-pagination");
+  }
+}
