+++ static/scripts/rewards/web3/erc721-permit.ts
@@ -1,12 +1,13 @@
 import { JsonRpcProvider, TransactionResponse } from "@ethersproject/providers";
 import { ethers } from "ethers";
 import { nftRewardAbi } from "../abis/nftRewardAbi";
-import { renderTransaction } from "../render-transaction/render-transaction";
+import { renderTransaction } from "../render-transaction/renderTransaction";
 import { Erc721Permit } from "../render-transaction/tx-type";
 import { claimButton, errorToast, loadingClaimButton, resetClaimButton, toaster } from "../toaster";
 import { connectWallet } from "./wallet";
+import { app } from "../app-state";
 
-export function claimErc721PermitHandler(permit: Erc721Permit, provider: JsonRpcProvider) {
+export function claimErc721PermitHandler(permit: Erc721Permit) {
   return async function claimButtonHandler() {
     const signer = await connectWallet();
     if (!signer) {
@@ -25,7 +26,7 @@ export function claimErc721PermitHandler(permit: Erc721Permit, provider: JsonRpc
       return;
     }
 
-    const isRedeemed = await isNonceRedeemed(permit, provider);
+    const isRedeemed = await isNonceRedeemed(permit, app.provider);
     if (isRedeemed) {
       toaster.create("error", `This NFT has already been redeemed.`);
       resetClaimButton();
@@ -44,7 +45,7 @@ export function claimErc721PermitHandler(permit: Erc721Permit, provider: JsonRpc
 
       claimButton.element.removeEventListener("click", claimButtonHandler);
 
-      renderTransaction(provider, true).catch((error) => {
+      renderTransaction(true).catch((error) => {
         console.error(error);
         toaster.create("error", `Error rendering transaction: ${error.message}`);
       });
